Fix MediaExcerpt creation
