+++ packages/browser-extension/public/options.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="options.js"></script>
+    <script type="module" src="options.mjs"></script>
   </body>
 </html>
