+++ packages/browser-extension/public/sidebar.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="sidebar.mjs"></script>
   </body>
 </html>
