+++ packages/browser-extension/src/content.ts
@@ -17,18 +17,26 @@ import {
   outcomeValence,
 } from "@sophistree/common";
 
-import "./highlights/outcome-colors.scss";
 import { AddMediaExcerptData } from "./store/entitiesSlice";
-import type {
-  ContentMessage,
-  CreateMediaExcerptMessage,
-  GetMediaExcerptsResponse,
-  MediaExcerptUpdates,
+import {
+  sidepanelKeepalivePortName,
+  type ContentMessage,
+  type CreateMediaExcerptMessage,
+  type GetMediaExcerptsResponse,
+  type MediaExcerptUpdates,
 } from "./extension/messages";
 import { deserializeMap } from "./extension/serialization";
 import { connectionErrorMessage } from "./extension/errorMessages";
 import * as contentLogger from "./logging/contentLogging";
 import { getPdfUrlFromViewerUrl, isPdfViewerUrl } from "./pdfs/pdfs";
+import {
+  AddMediaExcerptMessage,
+  GetMediaExcerptMessage,
+  GetMediaExcerptsMessage,
+  SelectMediaExcerptMessage,
+} from "./extension/chromeRuntimeMessages";
+
+import "./highlights/outcome-colors.scss";
 
 chrome.runtime.onConnect.addListener(getMediaExcerptsWhenSidebarConnects);
 chrome.runtime.onMessage.addListener(
@@ -56,8 +64,6 @@ function getMediaExcerptsWhenSidebarConnects(port: chrome.runtime.Port) {
     });
   }
 }
-
-export const sidepanelKeepalivePortName = "keepalive";
 
 async function handleMessage(
   message: ContentMessage,
@@ -426,45 +432,6 @@ async function selectMediaExcerpt(mediaExcerptId: string) {
   };
   await chrome.runtime.sendMessage(message);
 }
-
-interface AddMediaExcerptMessage {
-  action: "addMediaExcerpt";
-  data: AddMediaExcerptData;
-}
-
-interface SelectMediaExcerptMessage {
-  action: "selectMediaExcerpt";
-  data: {
-    mediaExcerptId: string;
-  };
-}
-
-interface GetMediaExcerptMessage {
-  action: "getMediaExcerpt";
-  data: {
-    mediaExcerptId: string;
-  };
-}
-
-interface GetMediaExcerptsMessage {
-  action: "getMediaExcerpts";
-  data: {
-    url: string;
-    canonicalUrl?: string;
-    pdfFingerprint?: string;
-  };
-}
-
-interface AuthStateChangedMessage {
-  action: "authStateChanged";
-}
-
-export type ChromeRuntimeMessage =
-  | AddMediaExcerptMessage
-  | SelectMediaExcerptMessage
-  | GetMediaExcerptMessage
-  | GetMediaExcerptsMessage
-  | AuthStateChangedMessage;
 
 declare global {
   interface Window {
