+++ packages/browser-extension/src/sidebar.tsx
